Enables cli indicated property file for session values
BegCli now requires apache commons cli 1.4 to receive the path to a
property file. Adds exception messages for BegCli as bc_messages...

BundlesEnoGettext exposes the expected property file ids. Able to
accept a property file.

Jre v10 isn't my default, so this adds a toolchain plugin to the maven
pom to direct it to that one. Adds version for source plugin. Specifies
copying resources as utf-8.