Enables emitting a key alias map
So I don't need to make these by hand or have an eighty character
literal to pick out the localized message. Creates the file in the
same package as properties or ListResourceBundle files. Adds another
template (or hardcoded if missing). Always prepares values for the
alias map, but only emits a file if the config indicates it should.

While I'm confident this ends this project, confidence isn't proof.
I'll test this another day with a toy project to retrieve the messages
with literal strings and the static alias references.