Valid example package; test script for locale files; map template
It turns out hyphen isn't a valid package character, so I substituted
underscore in the example property file.

I hadn't noticed that I'd forgotten the map template. Included.

Adds an example class to vet that these properties are valid and a
bash script to compile and run it. (As I had trouble figuring out
the right classpath and location.)