Uses messages.json instead of gettext files
Property for input is now input_path instead of input_dir. If
provided a directory, this will look for messages.json in
that directory. If provided a non json file, this will quit. Uses
alibaba's fastjson for this (though in a naive way, so it could
have been most any library).

Assumes json rather than gettext files. BundlesEnoGettext
is now misnamed. Emits as before, albiet with a time stamp
both for the file and the json it came from. EnoAlias
uses the canonical message ids as field names, rather
than category plus index. Creates the singular resource
bundle once, rather than appending to a file.